Fixed deleting message text after callback_data='givefamily'

+++ handlers/animal_handlers/allpet_handler.py
@@ -96,7 +96,7 @@ class AllpetHandler(BaseHandler):
             ]
             reply_markup = InlineKeyboardMarkup(keyboard)
 
-            if update.callback_query:
+            if update.callback_query and update.callback_query.data != 'givefamily':
                 await context.bot.delete_message(
                     chat_id=update.callback_query.message.chat_id,
                     message_id=update.callback_query.message.message_id
